Restructure to prepare for merge
