+++ python/spider-py/tests/client/test_driver.py
@@ -2,7 +2,7 @@
 
 import pytest
 
-from spider import chain, Driver, group, Int8, TaskContext
+from spider_py import chain, Driver, group, Int8, TaskContext
 
 MariaDBTestUrl = "jdbc:mariadb://127.0.0.1:3306/spider-storage?user=spider&password=password"
