--- python/src/spider/type/__init__.py
@@ -1,6 +1,0 @@
-"""Spider type package."""
-
-from spider.type.tdl_convert import to_tdl_type_str
-from spider.type.type import Double, Float, Int8, Int16, Int32, Int64
-
-__all__ = ["Double", "Float", "Int8", "Int16", "Int32", "Int64", "to_tdl_type_str"]
